requirements.txt
aioredis==2.0.1
alembic==1.10.4
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
autoflake==2.1.1
bcrypt==4.0.1
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.0.0
flower==1.2.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jwt==1.3.1
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.8.11
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pendulum==2.1.2
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prometheus-client==0.16.0
prometheus-fastapi-instrumentator==6.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
pyright==1.1.310
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
rsa==4.9
sentry-sdk==1.24.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.11.0
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
WTForms==3.0.1
